requirements.txt
Pillow==11.2.1
Requests==2.32.3
tqdm==4.66.5
